requirements.txt
cython==3.0.11
opencv-python==4.10.0.84
Pillow==10.2.0
mediapipe==0.10.20

# pip install torch==2.5.1 torchvision==0.20.1 -f https://download.pytorch.org/whl/cu118/torch_stable.html
torch==2.5.1
torchvision==0.20.1

# Additional dependencies for model downloading
requests>=2.25.0
tqdm>=4.60.0
onnxruntime-gpu>=1.12.0